Move built-in info utility function
